set fixed=false for all nodes
