+++ package.json
@@ -13,7 +13,7 @@
     },
     "dependencies": {
         "monaco-editor-webpack-plugin": "^3.0.1",
-        "nearley": "^2.19.2",
+        "nearley": "^2.20.1",
         "react": "^17.0.1",
         "react-dom": "^17.0.1",
         "react-monaco-editor": "^0.42.0",
