+++ src/components/AstPanel.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import "./AstPanel.scss";
-import { Ast } from "../syntax/parser";
+import type { Ast } from "../syntax/parser";
 import { FadeLoader } from "react-spinners";
 
 interface AstPanelProps {
