+++ src/components/PegPanel.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef, useState } from "react";
 import { Hypergraph } from "../analysis/hypergraph";
-import { Ast } from "../syntax/parser";
+import type { Ast } from "../syntax/parser";
 import { FadeLoader } from "react-spinners";
 
 import "./PegPanel.scss";
